[generic] Try parsing JWPlayer embedded videos (closes #12030)
