+++ ChangeLog
@@ -1,3 +1,9 @@
+version <unreleased>
+
+Extractors
++ [generic] Support complex JWPlayer embedded videos (#12030)
+
+
 version 2017.02.16
 
 Core
